+++ ControlManual/src/core/widgets/__init__.py
@@ -1,0 +1,2 @@
+from .console import Console
+from .feed import Feed
